Collect handler helpers into one file
